LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/events/uncompiled_event_handler_with_scripting_disabled-expected.txt

CONSOLE MESSAGE: line 1: ReferenceError: Can't find variable: this_will_error

FAIL when scripting is disabled, the handler is never compiled assert_equals: expected false but got true
